### 2025-06-27 d54614a
Add JOSS status badge
