### 2025-06-27 238bd7a
Add DOI for Monte Carlo SSA citation in bibliography
