+++ paper.bib
@@ -67,6 +67,7 @@
 	issn = {1520-0442},
 	shorttitle = {Monte {Carlo} {SSA}},
 	url = {https://journals.ametsoc.org/view/journals/clim/9/12/1520-0442_1996_009_3373_mcsdio_2_0_co_2.xml},
+	doi = {10.1175/1520-0442(1996)009<3373:MCSDIO>2.0.CO;2},
 	abstract = {Singular systems (or singular spectrum) analysis (SSA) was originally proposed for noise reduction in the analysis of experimental data and is now becoming widely used to identify intermittent or modulated oscillations in geophysical and climatic time series. Progress has been hindered by a lack of effective statistical tests to discriminate between potential oscillations and anything but the simplest form of noise, that is, “white” (independent, identically distributed) noise, in which power is independent of frequency. The authors show how the basic formalism of SSA provides a natural test for modulated oscillations against an arbitrary “colored noise” null hypothesis. This test, Monte Carlo SSA, is illustrated using synthetic data in three situations: (i) where there is prior knowledge of the power-spectral characteristics of the noise, a situation expected in some laboratory and engineering applications, or when the “noise” against which the data is being tested consists of the output of an independently specified model, such as a climate model; (ii) where a simple hypothetical noise model is tested, namely, that the data consists only of white or colored noise; and (iii) where a composite hypothetical noise model is tested, assuming some deterministic components have already been found in the data, such as a trend or annual cycle, and it needs to be established whether the remainder may be attributed to noise. The authors examine two historical temperature records and show that the strength of the evidence provided by SSA for interannual and interdecadal climate oscillations in such data has been considerably overestimated. In contrast, multiple inter- and subannual oscillatory components are identified in an extended Southern Oscillation index at a high significance level. The authors explore a number of variations on the Monte Carlo SSA algorithm and note that it is readily applicable to multivariate series, covering standard empirical orthogonal functions and multichannel SSA.},
 	language = {en},
 	urldate = {2024-12-21},
