### 2025-07-04 #8491 @abhishektiwari
@dadelforge Can you clarify contribution of Olivier de Viron, Marnik Vanclooster, and Niko Speybroeck to the software and paper? I only see [code contributions](https://github.com/ADSCIAN/ssalib/graphs/contributors) from 2 authors on the paper.
