### 2025-07-08 #8491 @dadelforge
Hi,
Thank you for handling the editorial and reviewing process. Regarding authorship, I declare that all authors have made significant contributions to this work. Although my co-authors are not advanced GitHub users and were not involved in pushing updates, they have made substantial contributions. Here are the defined contributions:
Software:
Original idea, conceptualization, and design patterns: DD, AA
Code writing and sharing: DD, AA, OdV
Code documentation, standardization, and packaging: DD
Validation, feedback, testing, and notebook tutorial development: DD, AA, OdV, MV, NS
Supervision: MV, OdV, NS
Funding acquisition: DD, AA, NS
Paper:
Writing the original draft: DD
Reviewing and editing: DD, AA, OdV, MV, NS
Regarding the story, SSALib has been a side project for a long time, starting during my PhD (2017-2020) with my colleague AA, under the supervision of MV. We learned about SSA through the R package rSSA and began developing Python code recipes for research and teaching purposes. I started working with OdV, who taught me about more advanced approaches, such as Monte Carlo SSA, and shared most of his Python code recipes, which I later integrated into the package. The project was on hold for a while. Last year, MV asked AA and I to teach SSA at a workshop, and NS and I started using SSA for a public health research project. We decided to resume this work and publish the package to support our future research and teaching activities.

Damien Delforge, PhD
Senior Researcher
Centre for Research on the Epidemiology of Disasters
Institute of Health and Society
University of Louvain (UCLouvain)

________________________________
De : Abhishek Tiwari ***@***.***>
Envoyé : vendredi 4 juillet 2025 21:47
À : openjournals/joss-reviews ***@***.***>
Cc : Damien Delforge ***@***.***>; Mention ***@***.***>
Objet : Re: [openjournals/joss-reviews] [PRE REVIEW]: SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis (Issue #8491)

[https://avatars.githubusercontent.com/u/363839?s=20&v=4]abhishektiwari left a comment (openjournals/joss-reviews#8491)<https://github.com/openjournals/joss-reviews/issues/8491#issuecomment-3037144752>

@dadelforge<https://github.com/dadelforge> Can you clarify contribution of Olivier de Viron, Marnik Vanclooster, and Niko Speybroeck to the software and paper? I only see code contributions<https://github.com/ADSCIAN/ssalib/graphs/contributors> from 2 authors on the paper.
