### 2025-07-08 #8491 @dadelforge
Here are some suggestions for reviewers:

  *
Wwhenxuan
  *
Tsbinns
  *
nmy2103

Please let me know if I should look for more profiles, inside or outside the JOSS reviewer database.
Best regards,
Damien

Damien Delforge, PhD
Senior Researcher
Centre for Research on the Epidemiology of Disasters
Institute of Health and Society
University of Louvain (UCLouvain)

________________________________
De : Abhishek Tiwari ***@***.***>
Envoyé : vendredi 4 juillet 2025 22:15
À : openjournals/joss-reviews ***@***.***>
Cc : Damien Delforge ***@***.***>; Mention ***@***.***>
Objet : Re: [openjournals/joss-reviews] [PRE REVIEW]: SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis (Issue #8491)

[https://avatars.githubusercontent.com/u/363839?s=20&v=4]abhishektiwari left a comment (openjournals/joss-reviews#8491)<https://github.com/openjournals/joss-reviews/issues/8491#issuecomment-3037183076>

@dadelforge<https://github.com/dadelforge> In addition, can you suggest a few potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our conflict of interest<https://github.com/openjournals/joss/blob/main/COI.md> (COI) when recommending reviewers.
