### 2025-07-09 #8491 @abhishektiwari
👋 @wwhenxuan @Tsbinns @nmy2103, would any of you be willing to review this submission for JOSS?  The submission under consideration is "SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis".

Based on your experience, we think you might be able to provide a great review of this submission.
Thanks
Abhi
