### 2025-07-16 #8491 @abhishektiwari
👋 @astrogilda @jwwthu, would any of you be willing to review this submission for JOSS? The submission under consideration is "SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis".

Based on your experience and interest in time series analysis and signal processing space, we think you might be able to provide a great review of this submission.
Thanks
Abhi
