### 2025-07-16 #8491 @wwhenxuan
> 👋 [@wwhenxuan](https://github.com/wwhenxuan) [@tsbinns](https://github.com/tsbinns) [@nmy2103](https://github.com/nmy2103), would any of you be willing to review this submission for JOSS? The submission under consideration is "SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis".
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
>
> Based on your experience, we think you might be able to provide a great review of this submission. Thanks Abhi

I'm sorry for the late reply 😥.
I just cloned SSALib, checked the SSA algorithm and ran all the examples. I have written SSA code myself, but I have to admit that the author not only provides richer and more efficient code, but also provides complete test examples. Through this project, I think even a scholar who has just come into contact with SSA can understand the basic principles of the algorithm well. Therefore, I think this work is very meaningful.
Here are my two suggestions:
1. I hope the author can provide more algorithm details in the paper, just like the test examples given.
2. Many of the current time series data are multivariate, so as a Python library for time series analysis, I hope the author can further add multivariate SSA to make SSALib more complete.

Finally, thank you for inviting me @abhishektiwari .
