### 2025-07-17 #8600 @editorialbot
[REVIEW]: SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis
**Submitting author:** @dadelforge (<a href="http://orcid.org/0000-0002-3552-9444">Damien Delforge</a>)
**Repository:** https://github.com/ADSCIAN/ssalib
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
