### 2025-10-06 #8600 @abhishektiwari
Hi @pitmonticone @Saran-nns, would any of you be willing to review this submission for JOSS? The submission under consideration is "SSALib: a Python Library for Time Series Decomposition using Singular Spectrum Analysis".

Based on your experience, we think you might be able to provide a great review of this submission. Thanks Abhi
