### 2025-10-27 f683fdd
- Update publication date in metadata - Improve and expand paper sections (Statement of Needs, Mathematical Background) - Add missing references to bibliography (Golyandina 2020, Hassani 2007)
