+++ paper.bib
@@ -30,13 +30,48 @@
 	keywords = {time series, forecasting, singular value decomposition, Multivariate Singular Spectrum Analysis, signal extraction ., signal processing},
 }
 
+@article{golyandina_particularities_2020,
+	title = {Particularities and commonalities of singular spectrum analysis as a method of time series analysis and signal processing},
+	volume = {12},
+	copyright = {© 2020 Wiley Periodicals, Inc.},
+	issn = {1939-0068},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/wics.1487},
+	doi = {10.1002/wics.1487},
+	language = {en},
+	number = {4},
+	urldate = {2025-10-27},
+	journal = {WIREs Computational Statistics},
+	author = {Golyandina, Nina},
+	year = {2020},
+	keywords = {decomposition, forecasting, signal processing, singular spectrum analysis, time series},
+	pages = {e1487},
+}
+
+
+@article{hassani_singular_2007,
+	title = {Singular {Spectrum} {Analysis}: {Methodology} and {Comparison}},
+	volume = {5},
+	issn = {1680-743X, 1683-8602},
+	shorttitle = {Singular {Spectrum} {Analysis}},
+	url = {https://jds-online.org/journal/JDS/article/1027},
+	doi = {10.6339/JDS.2007.05(2).396},
+	language = {en},
+	number = {2},
+	urldate = {2025-10-27},
+	journal = {Journal of Data Science},
+	author = {Hassani, Hossein},
+	month = aug,
+	year = {2007},
+	pages = {239--257},
+}
+
+
 @article{vautard_singular_1989,
 	title = {Singular spectrum analysis in nonlinear dynamics, with applications to paleoclimatic time series},
 	volume = {35},
 	issn = {0167-2789},
 	url = {https://www.sciencedirect.com/science/article/pii/0167278989900778},
 	doi = {10.1016/0167-2789(89)90077-8},
-	abstract = {We distinguish between two dimensions of a dynamical system given by experimental time series. Statistical dimension gives a theoretical upper bound for the minimal number of degrees of freedom required to describe tje attractor up to the accuracy of the data, taking into account sampling and noise problems. The dynamical dimension is the intrinsic dimension of the attractor and does not depend on the quality of the data. Singular Spectrum Analysis (SSA) provides estimates of the statistical dimension. SSA also describes the main physical phenomena reflected by the data. It gives adaptive spectral filters associated with the dominant oscillations of the system and clarifies the noise characteristics of the data. We apply SSA to four paleoclimatic records. The principal climatic oscillations, and the regime changes in their amplitude are detected. About 10 degrees of freedom are statistically significant in the data. Large noise and insufficient sample length do not allow reliable estimates of the dynamical dimension.},
 	number = {3},
 	urldate = {2024-12-21},
 	journal = {Physica D: Nonlinear Phenomena},
@@ -52,7 +87,6 @@
 	issn = {0167-2789},
 	url = {https://www.sciencedirect.com/science/article/pii/016727898690031X},
 	doi = {10.1016/0167-2789(86)90031-X},
-	abstract = {We consider the notion of qualitative information and the practicalities of extracting it from experimental data. Our approach, based on a theorem of Takens, draws on ideas from the generalized theory of information known as singular system analysis due to Bertero, Pike and co-workers. We illustrate our technique with numerical data from the chaotic regime of the Lorenz model.},
 	number = {2},
 	urldate = {2024-12-21},
 	journal = {Physica D: Nonlinear Phenomena},
@@ -68,7 +102,6 @@
[…]
