### 2025-10-27 32fb65e
Fix reference formatting in the Mathematical Background section
