### 2025-11-03 36798e6
Release version 0.1.2 (stable) and update GitHub Actions for publish job configuration.
