+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ssalib"
-version = "0.1.2b1"
+version = "0.1.2"
 description = "Singular Spectrum Analysis Library (SSALib)"
 readme = { file = "README.md", content-type = "text/markdown" }
 requires-python = ">=3.9"
