### 2025-11-03 efe8cbe
Update version in citation example to 0.1.2 and remove beta-specific note
