### 2025-11-03 39be496
Mark project as stable by updating classifiers in pyproject.toml and setup.py
