+++ pyproject.toml
@@ -19,7 +19,7 @@ authors = [
 urls = { homepage = "https://github.com/ADSCIAN/ssalib" }
 
 classifiers = [
-    "Development Status :: 4 - Beta",
+    "Development Status :: 5 - Production/Stable",
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: BSD License",
     "Programming Language :: Python :: 3",
