### 2025-11-03 2e4e9f3
Fix typo in GitHub Actions workflow: update Zenodo action reference.
