### 2025-11-03 6f5bfa3
Allow manual trigger for Zenodo publish workflow
