+++ .github/workflows/publish-zenodo.yml
@@ -3,6 +3,7 @@ name: Publish to Zenodo
 on:
   release:
     types: [published]
+  workflow_dispatch: {}
 
 jobs:
   publish:
