### 2025-11-03 f1b307e
Release version 0.1.3 and remove Zenodo publish workflow.
