+++ ssalib/__init__.py
@@ -3,7 +3,7 @@
 from .ssa import SingularSpectrumAnalysis
 from .montecarlo_ssa import MonteCarloSSA
 
-__version__ = '0.1.2'
+__version__ = '0.1.3'
 __author__ = ('Damien Delforge <damien.delforge@adscian.be>, '
               'Alice Alonso <alice.alonso@adscian.be>, '
               'Olivier de Viron, '
