+++ README.md
@@ -171,7 +171,7 @@ You can refer to SSALib using:
   author    = {Delforge, Damien AND Alonso, Alice AND de Viron, Olivier AND Vanclooster, Marnik AND Speybroeck, Niko},
   title     = {{SSALib}: A {Python} Library for {Singular Spectrum Analysis}},
   year      = {2025},
-  version   = {0.1.2},
+  version   = {0.1.3},
   url       = {https://github.com/ADSCIAN/ssalib}
 }
 ```
