+++ .github/workflows/publish-zenodo.yml
@@ -1,26 +1,0 @@
-name: Publish to Zenodo
-
-on:
-  release:
-    types: [published]
-  workflow_dispatch: {}
-
-jobs:
-  publish:
-    if: ${{ !github.event.release.prerelease }}  # Skip pre-releases
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - name: Publish to Zenodo
-        uses: zenodraft/action@v1
-        env:
-          ZENODO_TOKEN: ${{ secrets.ZENODO_TOKEN }}
-        with:
-          collection: ${{ secrets.ZENODO_COLLECTION_ID }} # Optional
-          draft: false
-          publish: true
-          sandbox: true # Set to true for testing
