### 2025-11-12 9f3d4b2
Fix typos, improve phrasing, and standardize formatting in paper.md and references.
