+++ paper.bib
@@ -128,7 +128,7 @@
 }
 
 @article{faouzi_pyts_2020,
-	title = {pyts: {A} {Python} {Package} for {Time} {Series} {Classification}},
+	title = {{pyts: A Python Package for Time Series Classification}},,
 	volume = {21},
 	url = {http://jmlr.org/papers/v21/19-763.html},
 	number = {46},
@@ -189,7 +189,7 @@
 }
 
 @inproceedings{seabold_statsmodels_2010,
-	title = {statsmodels: {Econometric} and statistical modeling with python},
+	title = {{Statsmodels: Econometric and Statistical Modeling with Python}},,
 	booktitle = {9th {Python} in {Science} {Conference}},
 	author = {Seabold, Skipper and Perktold, Josef},
 	year = {2010},
