### 2025-11-12 9430de0
Fix typo in reference entry by removing duplicate comma in paper.bib
