+++ paper.bib
@@ -128,7 +128,7 @@
 }
 
 @article{faouzi_pyts_2020,
-	title = {{pyts: A Python Package for Time Series Classification}},,
+	title = {{pyts: A Python Package for Time Series Classification}},
 	volume = {21},
 	url = {http://jmlr.org/papers/v21/19-763.html},
 	number = {46},
