### 2025-11-12 10f5077
Fix typo in reference entry by removing duplicate comma in paper.bib
