+++ paper.bib
@@ -189,7 +189,7 @@
 }
 
 @inproceedings{seabold_statsmodels_2010,
-	title = {{Statsmodels: Econometric and Statistical Modeling with Python}},,
+	title = {{Statsmodels: Econometric and Statistical Modeling with Python}},
 	booktitle = {9th {Python} in {Science} {Conference}},
 	author = {Seabold, Skipper and Perktold, Josef},
 	year = {2010},
