### 2025-11-12 910f25b
Fix typos and standardize AR abbreviation usage in paper.md
