### 2025-11-12 #8600 @dadelforge
Dear @abhishektiwari,

I apologize for the delay; I recently took a few days off. Thank you for your valuable additional feedback. I have now submitted an updated version of the paper and bibliography. All comments have been addressed, mainly by following suggestions, except I decided to keep "time-delayed" instead of "time-delay".

Please let me know how to proceed next.
