### 2025-11-14 #8600 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
