### 2025-11-14 ba243a6
minor changes in paper.md
