### 2025-11-14 #8600 @danielskatz
👋 @dadelforge - as track editor, I've proofread the paper, and suggest the changes in https://github.com/ADSCIAN/ssalib/pull/13 - please merge this or let me know what you disagree with, then we can proceed.
